Added functionality and tested the Unitree GO2
The following commit address the the issue that due to control heirarchy in Unitree Go2, the previously implemented archietcture of publishing values to the /wirellesscontroller topic may not always result in the robot movement, even after successful publication to the the respective topic. This commit adds a ROS workspace - go2_ws which exposes all the sports sdk commands as ROS2 services which can be used the ROS-MCP server. The following implementaion uses a variety of movements, acctions available at high level for the unitree go2. The given code has been tested on a real robot. In addition, for the setup in the README I have added instructioons o how to connect to the robot without the access to the WiFi router, it assumes that not everyone may have admin access.